+++ client/src/Pages/SignIn/index.css
@@ -14,6 +14,7 @@
     display: flex;
     flex-direction: column;
     justify-content: center;
+    background: azure;
   }
   .login .formContainer h1 {
     font-size: 25px;
